requirements.txt
asyncpraw==7.7.*
aiohttp==3.9.*
redis==5.*
celery==5.4.*
fastapi==0.111.*
uvicorn[standard]==0.29.*
psycopg[binary]==3.1.*
transformers==4.42.*
torch==2.3.*
pandas==2.2.*
python-dotenv==1.0.*
APScheduler==3.10.*
tqdm==4.66.*
